put common c fcns in common.c
