--- src/fallback/max.c
@@ -1,1 +1,0 @@
-../max.c
