+++ src/filesystem.cpp
@@ -22,38 +22,6 @@ namespace fs = std::filesystem;
 #include "filesystem.h"
 
 
-bool is_macos(){
-#if __APPLE__
-#include "TargetConditionals.h"
-#if TARGET_OS_MAC
-  return true;
-#endif
-#endif
-return false;
-}
-
-bool is_linux() {
-#ifdef __linux__
-  return true;
-#endif
-return false;
-}
-
-bool is_unix() {
-#ifdef __unix__
-  return true;
-#endif
-return false;
-}
-
-bool is_windows() {
-#ifdef _WIN32
-  return true;
-#endif
-return false;
-}
-
-
 size_t normal(const char* path, char* normalized) {
   // normalize path
   fs::path p(path);
@@ -70,13 +38,6 @@ size_t normal(const char* path, char* normalized) {
   return strlen(normalized);
 }
 
-
-bool sys_posix() {
-  char sep[2];
-
-  filesep(sep);
-  return sep[0] == '/';
-}
 
 void filesep(char* sep) {
   fs::path p("/");
