--- src/max.c
@@ -1,3 +1,0 @@
-#include "filesystem.h"
-
-size_t get_maxp(){ return MAXP; }
